requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain_community
langchain-huggingface
chromadb
pypdf
sentence-transformers